プレイリストの上書きと変換機能を追加

+++ lists.py
@@ -33,17 +33,19 @@ class listBase():
 
 	# 全消去
 	def deleteAllFiles(self):
-		for i in range(len(self.lst)):
-			self.deleteFile(i)
+		if self.deleteFile(0) != None:
+			self.deleteAllFiles()
 		self.playIndex = -1
 
 	def deleteFile(self, index):
+		rtn = None #戻り値の初期化
 		if index >= 0 and index < self.playIndex:
 			self.playIndex -= 1
 			rtn = self.lst.pop(index)
 			self.lcObject.DeleteItem(index)
 		# 再生中のファイルを削除したときのフラグ処理
-		elif index == self.playIndex:
+		elif index == self.playIndex and len(self.lst) != 0:
+			self.playIndex -= 1
 			isDeletePlayingFile =1
 			rtn = self.lst.pop(index)
 			self.lcObject.DeleteItem(index)
@@ -54,21 +56,17 @@ class listBase():
 
 	# 任意のファイルを取得（インデックス=最終getFile, 再生位置移動=しない）
 	def getFile(self, index=-2, movePos=False):
-		#インデックス指定の呼び出しに削除後の処理は考慮しない
+		#削除フラグ解除
+		self. isDeletePlayingFile = 0
 		if index >= 0 and index < len(self.lst):
-			self. isDeletePlayingFile = 0
 			rtn = self.lst[index]
 		elif index == -1 and len(self.lst) != 0:
-			self. isDeletePlayingFile = 0
 			index = len(self.lst)-1
 			rtn = self.lst[index]
 		# 第2引数なし（-2）で、再生中のファイルを返す
 		elif index == -2 and self.playIndex >= 0 and self.playIndex < len(self.lst):
-			if self.isDeletePlayingFile == 1:
-				rtn = ((None, None))((None, None))
-			else:
-				rtn = self.lst[self.playIndex]
-		# 再生位置リセット状態でも None を返す
+			rtn = self.lst[self.playIndex]
+		# 再生位置リセット状態では None を返す
 		else:
 			rtn = ((None, None))
 		if rtn[0] != None and index >= 0 and movePos == True:
@@ -88,13 +86,13 @@ class listBase():
 				self.playIndex = 0
 				return ((None, None))
 			return self.lst[self.playIndex]
+		elif self.playIndex == -1 and len(self.lst) != 0:
+			return self.lst[0]
 		else:
 			return ((None, None))
 
 class playlist(listBase):
 	def getNext(self):
-		if self.isDeletePlayingFile == 1:
-			return self.getFile()
 		if self.playIndex < len(self.lst)-1:
 			self.playIndex += 1
 			return self.lst[self.playIndex]
